almost 100% working code

+++ libflow.c
@@ -6,7 +6,7 @@
 /*   By: ocarlos- <ocarlos-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/30 21:42:14 by ocarlos-          #+#    #+#             */
-/*   Updated: 2020/07/29 16:28:25 by ocarlos-         ###   ########.fr       */
+/*   Updated: 2020/07/29 20:35:16 by ocarlos-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,7 @@ int		ft_vprintf(t_prtform *form, va_list lst, int count)
 {
 	char			*strarg;
 	int				argint;
-	unsigned long long int	argllint;
-	//unsigned int	argllint;
+	unsigned int	argllint;
 
 	argllint = 0;
 	if (form->typ == '%')
@@ -29,8 +28,7 @@ int		ft_vprintf(t_prtform *form, va_list lst, int count)
 	}
 	if (ft_testllint(form->typ))
 	{
-		argllint = va_arg(lst, unsigned long long int);
-		//argllint = va_arg(lst, unsigned int);
+		argllint = va_arg(lst, unsigned int);
 		form->siz = ft_redllint(argllint, form);
 	}
 	if (form->typ == 's')
@@ -55,8 +53,7 @@ int		ft_redint(int n, t_prtform *form)
 	return (form->siz);
 }
 
-int		ft_redllint(unsigned long long int n, t_prtform *form)
-//int		ft_redllint(unsigned int n, t_prtform *form)
+int		ft_redllint(unsigned int n, t_prtform *form)
 {
 	if (form->typ == 'u')
 		form->siz = ft_puin(n, form);
