+++ libutil.c
@@ -6,7 +6,7 @@
 /*   By: ocarlos- <ocarlos-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/30 21:37:55 by ocarlos-          #+#    #+#             */
-/*   Updated: 2020/07/21 16:58:32 by ocarlos-         ###   ########.fr       */
+/*   Updated: 2020/07/29 20:35:40 by ocarlos-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,8 +28,7 @@ t_prtform	*ft_initform(t_prtform *form)
 	return (form);
 }
 
-int			ft_countdig(unsigned long long int n, int b)
-//int			ft_countdig(unsigned int n, int b)
+int			ft_countdig(unsigned int n, int b)
 {
 	int	i;
 
@@ -57,8 +56,7 @@ void		ft_revstr(char *s, int size)
 	}
 }
 
-char		*ft_itoabase(unsigned long long int n, int b)
-//char		*ft_itoabase(unsigned int n, int b)
+char		*ft_itoabase(unsigned int n, int b)
 {
 	int		i;
 	int		r;
